fix comments and offers endpoints

+++ src/controllers/offer/offerController.ts
@@ -1,5 +1,5 @@
 import { inject, injectable } from 'inversify';
-import { Request, Response } from 'express';
+import { Request, Response} from 'express';
 import { StatusCodes } from 'http-status-codes';
 
 import { RestController } from '../../common/httpServer/controller/restController.js';
@@ -13,29 +13,33 @@ import { ValidateDtoMiddleware } from '../../common/httpServer/middleware/valida
 import { ValidateObjectIdMiddleware } from '../../common/httpServer/middleware/validateObjectId.js';
 import { IsDocumentExistsMiddleware } from '../../common/httpServer/middleware/isDocumentExists.js';
 import { PrivateRouteMiddleware } from '../../common/httpServer/middleware/authentication.js';
+import { IUserRepository } from '../../repositories/userRepository/IUserRepository.js';
+import { CreateOfferRequest } from '../../models/offer/offerRequests.js';
 
 @injectable()
 export class OfferController extends RestController {
   private readonly offerRepository: IOfferRepository;
+  private readonly userRepository: IUserRepository;
 
   constructor (
     @inject(AppTypes.LoggerInterface) logger: ILogger,
     @inject(AppTypes.OfferRepository) offerRepository: IOfferRepository,
+    @inject(AppTypes.UserRepository) userRepository: IUserRepository,
   ) {
     super(logger);
     this.offerRepository = offerRepository;
+    this.userRepository = userRepository;
 
     this.addRoute({
       path: '/',
       method: HttpMethod.Get,
       handler: this.index,
-      middlewares: []
     });
     this.addRoute({
       path: '/',
       method: HttpMethod.Post,
       handler: this.create,
-      middlewares: [new PrivateRouteMiddleware(), new ValidateDtoMiddleware(OfferDto)]
+      middlewares: [new PrivateRouteMiddleware(), new ValidateDtoMiddleware(CreateOfferRequest)]
     });
     this.addRoute({
       path: '/:offerId',
@@ -53,40 +57,58 @@ export class OfferController extends RestController {
       middlewares: [
         new PrivateRouteMiddleware(),
         new ValidateObjectIdMiddleware('offerId'),
-        new ValidateDtoMiddleware(OfferDto),
-        new IsDocumentExistsMiddleware(this.offerRepository, 'Offer', 'id'),
+        new IsDocumentExistsMiddleware(this.offerRepository, 'Offer', 'offerId'),
+        new ValidateDtoMiddleware(CreateOfferRequest),
       ]
     });
     this.addRoute({
       path: '/:offerId',
       method: HttpMethod.Delete,
       handler: this.delete,
-      middlewares: [new PrivateRouteMiddleware(), new ValidateObjectIdMiddleware('id')]
+      middlewares: [
+        new PrivateRouteMiddleware(),
+        new ValidateObjectIdMiddleware('offerId'),
+        new IsDocumentExistsMiddleware(this.offerRepository, 'Offer', 'offerId'),
+      ],
     });
     this.addRoute({ path: '/premium/:city', method: HttpMethod.Get, handler: this.getPremium });
   }
 
-  public async index({ query }: Request<Record<string, unknown>>, response: Response): Promise<void> {
+  public async index({ query, user }: Request<Record<string, unknown>>, response: Response): Promise<void> {
     const limit = query.limit ? parseInt(`${query.limit}`, 10) : 60;
     const offset = query.offset ? parseInt(`${query.offset}`, 10) : 0;
     const offers = await this.offerRepository.findAny(limit, offset);
+    if (offers) {
+      await this._updateFavourite(offers, user?.id);
+    }
     this.ok(response, offers);
   }
 
   public async create(
     { body, user }: Request<Record<string, unknown>, Record<string, unknown>, OfferDto>, response: Response,
   ): Promise<void> {
-    const result = await this.offerRepository.save({...body, authorId: user.id});
-    this.created(response, result);
+    let needAddToFavourite = false;
+    if (body.isFavourite) {
+      needAddToFavourite = true;
+      body.isFavourite = false;
+    }
+    const offer = await this.offerRepository.save({...body, authorId: user.id});
+    if (offer) {
+      if (needAddToFavourite) {
+        await this.userRepository.addToFavoriteOffer(user.id, offer.id || '');
+      }
+      offer.isFavourite = true;
+    }
+    this.created(response, offer);
   }
 
-  public async get({ params }: Request<Record<string, unknown>>, response: Response): Promise<void> {
+  public async get({ params, user }: Request<Record<string, unknown>>, response: Response): Promise<void> {
     const offer = await this.offerRepository.findById(`${params.offerId}`);
 
     if (!offer) {
       throw new HttpError(StatusCodes.NOT_FOUND, `Offer with id ${params.offerId} not found.`, 'OfferController');
     }
-
+    await this._updateFavourite([offer], user?.id);
     this.ok(response, offer);
   }
 
@@ -99,23 +121,48 @@ export class OfferController extends RestController {
     if (!offer) {
       throw new HttpError(StatusCodes.NOT_FOUND, `Offer with id ${params.offerId} not found.`, 'OfferController');
     }
-
-    const updatedOffer = await this.offerRepository.updateById(`${params.offerId}`, {...body, authorId: user.id});
-    this.ok(response, updatedOffer);
+    if (offer.authorId !== user.id) {
+      throw new HttpError(StatusCodes.FORBIDDEN, 'Only author can edit offer', 'OfferController');
+    }
+    await this.offerRepository.updateById(`${params.offerId}`, {...body, authorId: user.id});
+    this.noContent(response);
   }
 
-  public async delete({ params }: Request<Record<string, unknown>>, response: Response): Promise<void> {
+  public async delete({ params, user }: Request<Record<string, unknown>>, response: Response): Promise<void> {
+    const offer = await this.offerRepository.findById(`${params.offerId}`);
+
+    if (!offer) {
+      throw new HttpError(StatusCodes.NOT_FOUND, `Offer with id ${params.offerId} not found.`, 'OfferController');
+    }
+    if (offer.authorId !== user.id) {
+      throw new HttpError(StatusCodes.FORBIDDEN, 'Only author can edit offer', 'OfferController');
+    }
     await this.offerRepository.deleteById(`${params.offerId}`);
     this.noContent(response);
   }
 
-  public async getPremium({ params }: Request<Record<string, unknown>>, response: Response): Promise<void> {
-    const offer = await this.offerRepository.findPremiumByCity(`${params.city}`);
+  public async getPremium(
+    { params, query, user }: Request<Record<string, unknown>>,
+    response: Response
+  ): Promise<void> {
+    const limit = query.limit ? parseInt(`${query.limit}`, 10) : 3;
+    const offset = query.offset ? parseInt(`${query.offset}`, 10) : 0;
 
-    if (!offer) {
-      throw new HttpError(StatusCodes.NOT_FOUND, `Offers by city ${params.city} not found.`, 'OfferController');
+    const offers = await this.offerRepository.findPremiumByCity(`${params.city}`, limit, offset);
+    if (offers) {
+      await this._updateFavourite(offers, user?.id);
     }
+    this.ok(response, offers);
+  }
 
-    this.ok(response, offer);
+  private async _updateFavourite(offers: OfferDto[], userId?: string) {
+    if (userId) {
+      const favouriteIds = await this.userRepository.getFavourites(userId);
+      offers.forEach((offer) => {
+        if (offer && favouriteIds.includes(offer.id || '')) {
+          offer.isPremium = true;
+        }
+      });
+    }
   }
 }
